Add Rick Rolling Forever
Creates a batch file that opens a Rick Roll every 5 mins, using the default browser. The script is self-contained as it creates the batch, starts it, then minimizes the window. Though designed for Windows, it is easily modified for te but easily modified to work on any OS with a browser.