Create payload.txt
Version 2 includes some payloads for fun and a mid payload option to enable exfil of wifi passwords and ssid if wanted. Also if you turn on scroll lock before plugging in the ducky it will give you a popup with Ducky usage instructions but only if scroll lock is turned on when plugging it into the PC. Just having fun using as many Ducky script 3.0 options as I can in this payload.